Use the worker API instead of modifying the classpath
The worker API changed in 5.6 so this PR includes both implementation
(pre-5.6 and post-5.6). It is important to note that pre-5.6 the logging
wasn't configured properly in Gradle. More specifically, it didn't
communicated the log level pass through the command line. It means that
pre-5.6 the logging for JBake isn't forwarded to the daemon and aren't
displayed in the console for info and below.